prevent gcc crash on systems w 2GB RAM, by splitting cgalutils
