+++ src/cgalutils.h
@@ -11,6 +11,14 @@ typedef CGAL::Point_3<K> Vertex3K;
 typedef std::vector<Vertex3K> PolygonK;
 typedef std::vector<PolygonK> PolyholeK;
 
+namespace /* anonymous */ {
+        template<typename Result, typename V>
+        Result vector_convert(V const& v) {
+                return Result(CGAL::to_double(v[0]),CGAL::to_double(v[1]),CGAL::to_double(v[2]));
+       	}
+}
+
+
 namespace CGALUtils {
 	bool applyHull(const Geometry::ChildList &children, PolySet &P);
 	CGAL_Nef_polyhedron *applyOperator(const Geometry::ChildList &children, OpenSCADOperator op);
